KnightBoard.java a wild unresolved glitch has appeared!
My KnightBoard works until it doesn't.
It is able to recursively backtrack exactly 7 times before crashing. It is able to backtrack multiple steps, and if it finds a solution before backtracking 7 times it works out, but once it hits 7, the entire thing crashes down and burns. I've tried tracing through my code with pen and paper, but I can't find what's wrong with it. It seems to work... until it doesn't. This crazy bug is crazy.